[Test] Refactor test code
